Revert "minor changes"
This reverts commit 1b5a057f231dbe2109ccadc9a79c32e82e44c211.

+++ src/Modify.h
@@ -9,7 +9,7 @@ extern "C" void WINAPI get_file_name();
 extern "C" void WINAPI modify_source();
 
 extern "C" PatternScanner::ModuleInfo ZipScan;
-extern "C" std::uintptr_t file_name_rcx;
-extern "C" std::uintptr_t ret_addr_file_name;
-extern "C" std::uintptr_t ret_addr_file_source;
+extern "C" std::uint64_t file_name_rcx;
+extern "C" std::uint64_t ret_addr_file_name;
+extern "C" std::uint64_t ret_addr_file_source;
 #endif
